requirements.txt
streamlit
tensorflow
numpy
torch
openai-whisper
ffmpeg-python
regex
